Refactored code. Removed code to get information about users.
