+++ filterwords.js
@@ -1,0 +1,23 @@
+/* eslint-env node */
+
+module.exports = [
+  'ich', 'du', 'er', 'sie', 'es', 'wir', 'der', 'die', 'das', 'dem', 'ja', 'nein', 'zu', 'auch', 'kein', 'keine',
+  'ist', 'hat', 'auf', 'nicht', 'von', 'ein', 'eine', 'einen', 'und', 'dass', 'aber', 'kann', 'dazu', 'dann', 'noch',
+  'wenn', 'habe', 'mehr', 'nur', 'dir', 'mir', 'in', 'den', 'bitte', 'nach', 'werden', 'wird', 'machen', 'bzw', 'mit',
+  'ihr', 'da', 'jetzt', 'bald', 'alles', 'danke', 'ohne', 'innerhalb', 'also', 'mehr', 'schon', 'wo', 'wer', 'wie',
+  'warum', 'am', 'so', 'dich', 'gibt', 'diesen', 'bereits', 'oder', 'hier', 'um', 'im', 'an', 'sich', 'dort', 'bei',
+  'was', 'deine', 'denn', 'hier', 'für', 'man', 'wieder', 'haben', 'sind', 'mal', 'uns', 'über', 'hallo', 'aus', 'des',
+  'euch', 'gerade', 'geht', 'muss', 'sein', 'zur', 'können', 'zum', 'mich', 'alle', 'ok', 'als', 'gerne', 'kannst',
+  'war', 'neu', 'nun', 'hab', 'jemand', 'würde', 'nochmal', 'immer', 'habt', 'kommt', 'wäre', 'unsere', 'sollte',
+  'wurde', 'woher', 'dank', 'vielen', 'hi', 'benötigt', 'leider', 'unser', 'gleich', 'gleiche', 'soll', 'andere',
+  'anderen', 'einem', 'allerdings', 'ob', 'hast', 'diese', 'vielen', 'anscheinend', 'durch', 'kurz', 'mittlerweile',
+  'deshalb', 'darum', 'dafür', 'hätte', 'würdest', 'sehr', 'zusammen', 'erreichen', 'fehlt', 'irgendwie', 'brauchen',
+  'vom', 'dran', 'bin', 'etwas', 'beim', 'könnte', 'morgen', 'heute', 'gestern', 'damit', 'schreiben', 'zugreifen',
+  'bzgl', 'doch', 'evtl', 'gut', 'will', 'benötigen', 'dies', 'klar', 'kleiner', 'könnt', 'lassen', 'gesagt', 'wurden',
+  'vorher', 'betrifft', 'erste', 'je', 'erstmal', 'weder', 'ggf', 'unter', 'scheint', 'zwischen', 'bis', 'läuft',
+  'nichts', 'läuft', 'hatte', 'vielleicht', 'weil', 'komme', 'kommen', 'falls', 'welche', 'später', 'seit', 'ihn',
+  'sonst', 'bist', 'bekommen', 'ganz', 'wohl', 'erst', 'sehen', 'davon', 'meine', 'glaube', 'neuen', 'besten',
+  'schau', 'schaue', 'viel', 'müsste', 'möchte', 'sollten', 'eigentlich', 'natürlich', 'unseren',
+  'you', 'the', 'and', 'to', 'for', 'true', 'false', 'is', 'we', 'not', 'no', 'sorry', 'can', 'hey', 'okay', 'that',
+  'are', 'this', 'at', 'of', 'please'
+]
